Redcuce jumping effect when new nodes are created.
Incorporated fork of Springy.js that places nodes close to their
parents when they are created, instead of creating new nodes at a
random location. Works well to reduce the jumping effect.

Added a bit of logic to hide the node's KineticJS shape until it
has been positioned by springy, to avoid the effect of the node
momentarily appearing at the origin.