Constrain velocity to prevent graph destabilization.
When node has a large number of edges, the Springy calculations tend to
cause the node to have erratic behavior, which in some cases can lead to
destabilization of the entire graph. Put a speed limit in place
(maxVelocity): when velocity is updated, if the velocity is greater than
maxVelocity, divide the velocity by a factor of 10. I have not been able
to reproduce graph destabilization with this in place.